pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genesis-ai"
version = "0.1.0"
description = "Genesis AI Evolution Laboratory - Evolutionary algorithms for efficient AI model creation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Genesis Team"}
]
keywords = [
    "machine-learning",
    "evolutionary-algorithms",
    "knowledge-distillation",
    "model-pruning",
    "deep-learning",
    "pytorch",
    "transformers",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    "ray[tune]>=2.9.0",
    "deepspeed>=0.12.0",
    "deap>=1.4.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "tensorboard>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
tts = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyworld>=0.3.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/genesis-ai/genesis"
Documentation = "https://genesis-ai.github.io/genesis"
Repository = "https://github.com/genesis-ai/genesis"
Issues = "https://github.com/genesis-ai/genesis/issues"

[project.scripts]
genesis = "genesis.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["genesis*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
